📊 Improve wave report tower display

+++ src/Bearded.TD/UI/Extensions.cs
@@ -29,7 +29,7 @@ static class Extensions
         this AnchorTemplate a, Control control, double? width = null, double margin = 0) =>
         a.Right(margin: control.HorizontalAnchors.Left.Offset + margin, width: width);
 
-    public static AnchorTemplate TopToBottomPercentage(this AnchorTemplate template, float top, float bottom)
+    public static AnchorTemplate TopToBottomPercentage(this AnchorTemplate template, double top, double bottom)
     {
         return template.Top(relativePercentage: top).Bottom(relativePercentage: bottom);
     }
